backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
redis==5.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
yt-dlp==2024.12.13
google-generativeai==0.6.0
rq==1.16.1
requests==2.31.0
pytz==2024.1
aiofiles==23.2.1
email-validator==2.1.0

google-cloud-speech==2.26.0
